sutt_project/requirements.txt
asgiref==3.11.0
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
cryptography
dj-database-url==3.1.0
Django==5.1.3
django-allauth==65.14.0
django-allauth-ui==1.8.1
django-markdownx==4.0.9
django-widget-tweaks==1.5.1
python-dotenv==1.2.1
gunicorn==25.0.3
idna==3.11
Markdown==3.10.1
martor==1.7.16
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
PyYAML==6.0.3
requests==2.32.5
slippers==0.6.2
sqlparse==0.5.5
whitenoise==6.11.0
